ui-updated

+++ app/src/main/java/com/ahoy/weatherapp/presentation/ui/main/CurrentWeatherUIState.kt
@@ -5,9 +5,9 @@ import com.google.gson.annotations.SerializedName
 
 data class CurrentWeatherUIState (
 
-  var temp       : Double?            = null,
+  var temp       : Int?            = null,
   var humidity : Int?               = null,
-  var wind       : Double?              = null,
+  var wind       : Int?              = null,
   var date       : String,
   var name       : String,
   var iconCode       : String,
